+++ src/route/app/dividends/(container)/calendar/(components)/CalendarItem.tsx
@@ -8,7 +8,7 @@ import { Dividend } from '@Types/index';
 import { getAllDividends } from '@Lib/api/dividend';
 import DayJs from '@Utils/DayJs';
 import Loader from '@Components/loader';
-import { div, mul, sub } from '@Utils/Math';
+import { add, div, mul, sub } from '@Utils/Math';
 
 interface CalendarItemProps {
   date: Dayjs;
@@ -16,6 +16,8 @@ interface CalendarItemProps {
 
 export default function CalendarItem({ date }: CalendarItemProps) {
   const [, setCalendarDate] = useAtom(AppStore.calendar.date);
+  const [, setCalendarDividend] = useAtom(AppStore.calendar.dividend);
+  const beforeSetId = useRef<boolean>(false);
 
   const { data, isSuccess, isFetching } = useQuery<Dividend[]>({
     queryKey: ['Dividends', date.format('YYYY-MM-DD')],
@@ -38,21 +40,64 @@ export default function CalendarItem({ date }: CalendarItemProps) {
     setHeight(Number(document.getElementById('test')?.clientHeight)),
   );
 
+  const handleStore = (e: React.MouseEvent<HTMLDivElement>) => {
+    if (e.currentTarget.classList.contains('cursor-wait')) return;
+
+    if (!beforeSetId.current) {
+      // false
+      if (e.currentTarget.id === 'group') {
+        // id === 'group' && false
+        handleStoreDate(e);
+      } else {
+        // id !== 'group' && false
+        handleStoreId(e);
+        beforeSetId.current = !0;
+      }
+    } else {
+      // true
+      if (e.currentTarget.id === 'group') {
+        // id === 'group' && true
+        beforeSetId.current = !1;
+      } else {
+        // id !== 'group' && true
+        handleStoreId(e);
+      }
+    }
+  };
+
   const handleStoreDate = (e: React.MouseEvent<HTMLDivElement>) => {
+    setCalendarDate(date.format('YYYY-MM-DD'));
+    if (isSuccess && !isFetching && data?.length) {
+      setCalendarDividend(
+        data.sort((a, b) =>
+          add(
+            sub(a.dividendPerShare, b.dividendPerShare),
+            sub(DayJs(a.dateExDividend).unix(), DayJs(b.dateExDividend).unix()),
+          ),
+        )[0].id,
+      );
+    } else {
+      setCalendarDividend('');
+    }
+  };
+
+  const handleStoreId = (e: React.MouseEvent<HTMLDivElement>) => {
     if (e.currentTarget.classList.contains('cursor-wait')) return;
     setCalendarDate(date.format('YYYY-MM-DD'));
+    setCalendarDividend(e.currentTarget.id);
   };
 
   return (
     <div
+      id="group"
       className={
-        'bg-slate-900 rounded-lg w-full h-full p-2 gap-y-2 flex flex-col justify-start transition-all duration-300 ease-in-out hover:z-10 z-0 ' +
+        'bg-slate-900 rounded-lg w-full h-full p-2 gap-y-2 flex flex-col justify-start transition-all duration-300 ease-in-out hover:z-10 z-10 ' +
         (isFetching ? 'cursor-wait' : 'cursor-pointer') +
         (DayJs().isSame(date, 'day')
           ? ' hover:ring-4 ring-2 ring-offset-2 ring-offset-slate-500/50 hover:ring-slate-900 ring-slate-800'
           : ' hover:ring-4 ring-offset-2 ring-offset-slate-800 ring-slate-900')
       }
-      onClick={handleStoreDate}
+      onClick={handleStore}
     >
       <div className="text-center text-lg">{date.date()}</div>
 
@@ -76,7 +121,9 @@ export default function CalendarItem({ date }: CalendarItemProps) {
               .map((dividend) => (
                 <div
                   key={dividend.id}
-                  className="bg-slate-800 rounded-lg h-fit text-sm text-center p-0.5"
+                  id={dividend.id}
+                  className="bg-slate-800 rounded-lg h-fit text-sm text-center p-0.5 z-0"
+                  onClick={handleStore}
                 >
                   {dividend.stockSymbol}
                 </div>